classwork/api-app/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
spotipy==2.23.0
requests==2.31.0
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
